# infra/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "高级 RAG 知识库独立复现",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "summary": "Register",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credentials"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/knowledge-bases": {
      "get": {
        "summary": "List Kbs",
        "operationId": "list_kbs_api_knowledge_bases_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "summary": "Create Kb",
        "operationId": "create_kb_api_knowledge_bases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KBCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/knowledge-bases/{kb_id}/documents": {
      "get": {
        "summary": "List Documents",
        "operationId": "list_documents_api_knowledge_bases__kb_id__documents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Upload Document",
        "operationId": "upload_document_api_knowledge_bases__kb_id__documents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "kb_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kb Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_knowledge_bases__kb_id__documents_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/retrieval/debug": {
      "post": {
        "summary": "Retrieval Debug",
        "operationId": "retrieval_debug_api_retrieval_debug_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/chat/runs": {
      "post": {
        "summary": "Submit",
        "operationId": "submit_api_chat_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/runs/{run_id}": {
      "get": {
        "summary": "Trace",
        "operationId": "trace_api_runs__run_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_document_api_knowledge_bases__kb_id__documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "tags": {
            "type": "string",
            "title": "Tags",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_knowledge_bases__kb_id__documents_post"
      },
      "ChatRequest": {
        "properties": {
          "kb_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 1,
            "title": "Kb Id"
          },
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "filters": {
            "$ref": "#/components/schemas/MetadataFilter"
          },
          "strategy": {
            "type": "string",
            "enum": [
              "dense",
              "hybrid",
              "full",
              "custom"
            ],
            "title": "Strategy",
            "default": "custom"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 36
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "mode": {
            "type": "string",
            "enum": [
              "agent",
              "rag"
            ],
            "title": "Mode",
            "default": "agent"
          },
          "application_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 36
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kb_id",
          "query"
        ],
        "title": "ChatRequest"
      },
...
      "Credentials": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "pattern": "^[a-zA-Z0-9_.-]+$",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Password"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Credentials"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "KBCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "title": "Description",
            "default": ""
          },
          "config": {
            "$ref": "#/components/schemas/RetrievalConfig"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "KBCreate"
      },
      "MetadataFilter": {
        "properties": {
          "document_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Document Ids"
          },
          "file_types": {
            "items": {
              "type": "string",
              "enum": [
                "pdf",
                "docx",
                "xlsx",
                "md",
                "txt"
              ]
            },
            "type": "array",
            "maxItems": 5,
            "title": "File Types"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Tags"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "MetadataFilter"
      },
...
      "QueryRequest": {
        "properties": {
          "kb_id": {
            "type": "string",
            "maxLength": 36,
            "minLength": 1,
            "title": "Kb Id"
          },
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query"
          },
          "filters": {
            "$ref": "#/components/schemas/MetadataFilter"
          },
          "strategy": {
            "type": "string",
            "enum": [
              "dense",
              "hybrid",
              "full",
              "custom"
            ],
            "title": "Strategy",
            "default": "custom"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kb_id",
          "query"
        ],
        "title": "QueryRequest"
      },
      "RetrievalConfig": {
        "properties": {
          "query_rewrite": {
            "type": "boolean",
            "title": "Query Rewrite",
            "default": false
          },
          "multi_query": {
            "type": "boolean",
            "title": "Multi Query",
            "default": false
          },
          "hyde": {
            "type": "boolean",
            "title": "Hyde",
            "default": false
          },
          "hybrid": {
            "type": "boolean",
            "title": "Hybrid",
            "default": true
          },
          "rerank": {
            "type": "boolean",
            "title": "Rerank",
            "default": false
          },
          "parent_retrieval": {
            "type": "boolean",
            "title": "Parent Retrieval",
            "default": true
          },
          "metadata_filter": {
            "type": "boolean",
            "title": "Metadata Filter",
            "default": true
          },
          "top_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 8
          },
          "candidate_k": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Candidate K",
            "default": 16
          },
          "context_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Context K",
            "default": 8
          },
          "rerank_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Rerank K",
            "default": 4
          },
          "rrf_k": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 1.0,
            "title": "Rrf K",
            "default": 60
          },
          "rerank_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Rerank Threshold",
            "default": 0.05
          },
          "cosine_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Cosine Threshold",
            "default": 0.3
          },
          "context_chars": {
            "type": "integer",
            "maximum": 30000.0,
            "minimum": 500.0,
            "title": "Context Chars",
            "default": 10000
          },
          "chunk_strategy": {
            "type": "string",
            "enum": [
              "recursive",
              "recursive_short",
              "parent_child"
            ],
            "title": "Chunk Strategy",
            "default": "parent_child"
          },
          "structure_mode": {
            "type": "string",
            "enum": [
              "general",
              "laws",
              "qa"
            ],
            "title": "Structure Mode",
            "default": "general"
          },
          "recursive_size": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 100.0,
            "title": "Recursive Size",
            "default": 500
          },
          "recursive_overlap": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 0.0,
            "title": "Recursive Overlap",
            "default": 50
          },
          "short_size": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 50.0,
            "title": "Short Size",
            "default": 250
          },
          "short_overlap": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 0.0,
            "title": "Short Overlap",
            "default": 30
          },
          "parent_size": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 200.0,
            "title": "Parent Size",
            "default": 1200
          },
          "child_size": {
            "type": "integer",
            "maximum": 1000.0,
            "minimum": 50.0,
            "title": "Child Size",
            "default": 300
          },
          "child_overlap": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 0.0,
            "title": "Child Overlap",
            "default": 0
          },
          "grade_retry_limit": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Grade Retry Limit",
            "default": 3
          },
          "check_retry_limit": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Check Retry Limit",
            "default": 2
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "RetrievalConfig"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}